Update for Windows with binaries
We provide the updates to make lispe editor work with a mouse.

This version is now on par with linux and Mac OS.
Windows executable are available in binaries/windows

+++ src/strings.cxx
@@ -158,7 +158,7 @@ public:
                 Element* ch;
                 List* result = new List;
                 wstring localvalue;
-                size_t pos = 0;
+                long pos = 0;
                 if (search_string == L"") {
                     long sz = strvalue.size();
                     //we split the string into an array of characters
@@ -198,7 +198,7 @@ public:
                 Element* ch;
                 List* result = new List;
                 wstring localvalue;
-                size_t pos = 0;
+                long pos = 0;
                 if (search_string == L"") {
                     long sz = strvalue.size();
                     //we split the string into an array of characters
